Enabled api testing
