+++ backend/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "jest --config=test/jest.config.json --runInBand",
-    "dev": "nodemon src/app.js",
-    "postman": "nodemon src/app.js auth=postman"
+    "dev": "nodemon src/server.js",
+    "postman": "nodemon src/server.js auth=postman"
   },
   "author": "Xander Palmer",
   "license": "MIT",
@@ -21,6 +21,7 @@
   "devDependencies": {
     "@types/jest": "^29.5.3",
     "jest": "^29.6.1",
-    "nodemon": "^3.0.1"
+    "nodemon": "^3.0.1",
+    "supertest": "^6.3.3"
   }
 }
